requirements.txt
asgiref==3.3.0
Django==3.0
django-rest-auth==0.9.5
djangorestframework==3.12.1
Pillow==8.0.0
pkg-resources==0.0.0
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
